feat: move runBatchesInParallel function to its own file
